Fix computation stack overflow check
The size of the computation stack was miscoded to overlap with the
Terminal Input Buffer.  This meant that if you pushed a lot of things
onto the stack, you would eventually collide with terminal input,
one overwriting the other.

Now, if you overflow the stack, you get a MSG #7 instead of a hang.

It would probably be better to recode this somehow to compute the
size using the start of the computation stack and the start of the
TIB.