Improve ?STACK (QSTK) command
QSTK checks for computation stack under/overflow.

The original version seems to not work right because if you did a
DOT (.) command on an empty stack, you would not get an error.

I think that this was caused by the QSTK command pushing a couple of
items onto the stack before reading the current stack pointer to use
to compare with the stack boundaries.

The fix is to read the current stack pointer first before pushing
anything else to make sure we are testing the pointer at the time
QSTK was invoked.